First GPU version, with research log
First cut at a Cuda GPU program for executing the main grid counting loop for the Sudoku 4x3 exact count. It makes very poor use of GPU resources, and is actually slower than running the CPU code on the Jetson's 8-core ARM v8.2 processors. The purpose of this first cut is to confirm that I understand the Nvidia tool chain and the most basic operations of Cuda. The code runs and gets the correct results.